Fixed conditional logic bug.
The conditional logic wasn't being passed through to the
customizer-admin-js script.

The problem was caused by the fact that the script localization was
being handled exclusively by the WPC_Extended plugin, and as such it
called an instance of the WPC_Extended class. For settings created using
the WPC_Extended_Sass class, the plugin needs to pass its own data to
the localization function.

The simple way to do this is by removing the action which enqueues the
script, and replacing it with a new one which references the correct
class!